+++ jest.config.js
@@ -1,0 +1,16 @@
+const config = {
+  extensionsToTreatAsEsm: ['.ts'],
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  transform: {
+    '^.+\\.[tj]sx?$': [
+      'ts-jest',
+      {
+        useESM: true,
+      },
+    ],
+  }
+};
+
+// eslint-disable-next-line import/no-unused-modules
+export default config;